MCP_GATEWAY/SKill_SOCKET-Mcp_gateway/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
httpx==0.27.2
python-dotenv==1.0.1
scikit-learn==1.5.2
numpy==1.26.4
pandas==2.2.3
